backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.26.0
ccxt>=4.0.0
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
pydantic[email]>=2.4.0
sqlalchemy>=1.4.0
aiosqlite>=0.17.0
asyncpg>=0.24.0
psycopg2-binary>=2.9.0
greenlet>=1.0.0
jinja2>=3.0.0
google-generativeai>=0.3.0
pywebpush>=1.14.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
python-multipart>=0.0.6
argon2-cffi>=21.0.0
ft-pandas-ta==0.3.16
redis>=4.0.0
slowapi>=0.1.9
python-telegram-bot>=20.0
stripe>=8.0.0
bcrypt>=4.0.1
